zdejmowanie produktów ze stanu, częściowe zabezpieczenie
produkty które są na stanie w koszyku ciągle się odejmują, gdy występuje błąd

+++ shop_app/models.py
@@ -2,7 +2,7 @@ from django.db import models
 
 class Category(models.Model):
     name = models.CharField(max_length=50, verbose_name="Nazwa kategorii")
-    
+
     def __str__(self):
         return self.name
 
@@ -15,7 +15,7 @@ class Product(models.Model):
     name = models.CharField(max_length=50, verbose_name="Nazwa")
     description = models.TextField(max_length=4000, blank=True, verbose_name="Opis")
     price = models.DecimalField(max_digits=6, decimal_places=2, verbose_name="Cena")
-    quantity = models.IntegerField(default=0, verbose_name="Ilość")
+    quantity = models.PositiveSmallIntegerField(default=0, verbose_name="Ilość")
     image = models.FileField(default="no_photo.png",verbose_name="Zdjęcie")
 
     def __str__(self):
